scripts/requirements.txt
pillow>=9.0.0,<9.1.0
ipython>=8.0.1
flake8>=4.0.1
black>=21.12b0